Moved logger to lib.log
